Add seaborn to dependency list
Due to the missing seaborn dependency mention, pip did not install
seaborn which lead to a ModuleNotFoundError during import time. This
will fix it if accepted.

Tested on a empty environment by installing using `pip setup.py
install` before and after the addition. The error was not raised
post-addition.

This commit should resolve #517.